+++ application/packages/shared/src/core/index.ts
@@ -1,2 +1,3 @@
 export * from './events/events.bus'
 export * from './events/events.types'
+export * from './events/events.schema'